Add default "error" texture.
If a texture that needs to be loaded is not found, load up a default
"error" texture. Currently, it seems like the "test.png" file is being
used like this within the egnine (see Material class). This will
standardize that practice.

This type of fix is used by Half Life and many other AAA games alike.